+++ mithril-common/src/chain_reader/pallas_chain_reader.rs
@@ -86,9 +86,7 @@ impl PallasChainReader {
             }
             NextResponse::RollBackward(rollback_point, _) => {
                 let chain_point = ChainPoint::from(rollback_point);
-                Ok(Some(ChainBlockNextAction::RollBackward {
-                    slot_number: chain_point.slot_number,
-                }))
+                Ok(Some(ChainBlockNextAction::RollBackward { chain_point }))
             }
             NextResponse::Await => Ok(None),
         }
@@ -280,8 +278,8 @@ mod tests {
         let (_, client_res) = tokio::join!(server, client);
         let chain_block = client_res.expect("Client failed to get next chain block");
         match chain_block {
-            ChainBlockNextAction::RollBackward { slot_number } => {
-                assert_eq!(slot_number, get_fake_chain_point_backwards().slot_number);
+            ChainBlockNextAction::RollBackward { chain_point } => {
+                assert_eq!(chain_point, get_fake_chain_point_backwards());
             }
             _ => panic!("Unexpected chain block action"),
         }